refactor: add methods and implement traits for `FilePath` and `SafeFilePath`
closes #1726
