+++ .changes/fs-dialog-file-path-methods.md
@@ -1,0 +1,10 @@
+---
+"fs": patch
+"dialog": patch
+---
+
+Add utility methods on `FilePath` and `SafeFilePath` enums which are:
+
+- `path`
+- `simplified`
+- `into_path`
