+++ .changes/fs-dialog-file-path-traits.md
@@ -1,0 +1,6 @@
+---
+"fs": patch
+"dialog": patch
+---
+
+Implement `Serialize`, `Deserialize`, `From`, `TryFrom` and `FromStr` traits for `FilePath` and `SafeFilePath` enums.
