+++ .changes/fs-dialog-non-exhaustive-error.md
@@ -1,0 +1,6 @@
+---
+"fs": patch
+"dialog": patch
+---
+
+Mark `Error` enum as `#[non_exhuastive]`.
