+++ .changes/fs-dialog-safe-file-path.md
@@ -1,0 +1,6 @@
+---
+"fs": patch
+"dialog": patch
+---
+
+Add `SafeFilePath` enum.
