+++ Cargo.toml
@@ -11,7 +11,7 @@ resolver = "2"
 [workspace.dependencies]
 serde = { version = "1", features = ["derive"] }
 log = "0.4"
-tauri = { version = "2.0.0-rc.8", default-features = false }
+tauri = { git = "https://github.com/tauri-apps/tauri", branch = "feat/serialize-safe-pathbuf", default-features = false }
 tauri-build = "2.0.0-rc.7"
 tauri-plugin = "2.0.0-rc.7"
 tauri-utils = "2.0.0-rc.7"
