+++ plugins/dialog/Cargo.toml
@@ -26,7 +26,6 @@ serde_json = { workspace = true }
 tauri = { workspace = true }
 log = { workspace = true }
 thiserror = { workspace = true }
-dunce = { workspace = true }
 url = { workspace = true }
 tauri-plugin-fs = { path = "../fs", version = "2.0.0-rc.1" }
