+++ plugins/fs/Cargo.toml
@@ -30,6 +30,7 @@ uuid = { version = "1", features = ["v4"] }
 glob = "0.3"
 notify = { version = "6", optional = true, features = ["serde"] }
 notify-debouncer-full = { version = "0.3", optional = true }
+dunce = { workspace = true }
 
 [features]
 watch = ["notify", "notify-debouncer-full"]
